+++ core/src/com/pacman/Pantallas/PantallaMenu.java
@@ -4,7 +4,6 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
-import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
@@ -14,7 +13,7 @@ import com.pacman.JuegoPrincipal;
 public class PantallaMenu extends PantallaBase {
 
     private TextButton jugar, rankings, salir;
-    private Table buttons;
+    private Table tabla;
     private Label cabecera;
 
     public PantallaMenu(JuegoPrincipal juego) {
@@ -24,10 +23,10 @@ public class PantallaMenu extends PantallaBase {
     @Override
     public void show() {
         super.show();
-        establecerCamara();
-        skin = new Skin(Gdx.files.internal("skin/neon-ui.json"));//skin para los botones
-        //se crea la cabecera
-        //Label.LabelStyle labelStyle=new Label.LabelStyle(,Color.ORANGE);
+        //se desabilita el boton de retroceso
+        retroceso.remove();
+
+        //se establece el titulo el juego
         cabecera = new Label("PAC MAN", skin);
         cabecera.setSize(100, 100);
         cabecera.setPosition(altoEnPx / 2, 300);
@@ -52,17 +51,24 @@ public class PantallaMenu extends PantallaBase {
         });
 
         salir = new TextButton(traductor.get("pantallaMenu.salir"), skin);
-        buttons = new Table();
-        buttons.setFillParent(true);                       //redimensiona el tamaño de la tabla al del stage
-        buttons.add(cabecera).height(50).width(200).padBottom(60);
-        buttons.row();
-        buttons.add(jugar).height(50).width(200).padBottom(30);
-        buttons.row();
-        buttons.add(rankings).height(50).width(200).padBottom(30);
-        buttons.row();
-        buttons.add(salir).height(50).width(200).padBottom(30);
+        salir.addListener(new ClickListener(){
+            @Override
+            public void clicked(InputEvent event, float x, float y) {
+                Gdx.app.exit();
+            }
+        });
+        tabla = new Table();
+        tabla.setFillParent(true);                       //redimensiona el tamaño de la tabla al del stage
+        tabla.add(cabecera).height(50).width(200).padBottom(60);
+        tabla.row();
+        tabla.add(jugar).height(50).width(200).padBottom(30);
+        tabla.row();
+        tabla.add(rankings).height(50).width(200).padBottom(30);
+        tabla.row();
+        tabla.add(salir).height(50).width(200).padBottom(30);
+
         Gdx.input.setInputProcessor(escenario);
-        escenario.addActor(buttons);
+        escenario.addActor(tabla);
     }
 
     @Override
